requirements.txt
asgiref==3.2.10
Babel==2.8.0
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.0.8
django-extensions==3.0.3
django-import-export==2.3.0
django-phonenumber-field==4.0.0
django-phonenumbers==1.0.1
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.4
phonenumbers==8.12.6
Pillow==7.2.0
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.46
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
xlrd==1.2.0
xlwt==1.3.0
